+++ package.json
@@ -4,7 +4,8 @@
   "description": "CS546 Final Project",
   "main": "app.js",
   "scripts": {
-    "start": "node app.js"
+    "start": "node app.js",
+    "seed":"node ./tasks/seed.js"
   },
   "repository": {
     "type": "git",
