+++ views/home/home.handlebars
@@ -1,24 +1,25 @@
 <section>
-        <nav class="topnav">
-            <a class="active" href="/">Home</a>
-            <a href="/customers/login">Customers</a>
-            <a href="/restaurants/login">Restaurant</a>
-            <a href="/reviews">Reviews</a>
-        </nav>
-        </section>
+    <nav class="topnav">
+        <a class="active" href="/">Home</a>
+        <a href="/customers/login">Customers</a>
+        <a href="/restaurants/login">Restaurant</a>
+        <a href="/reviews">Reviews</a>
+    </nav>
+</section>
 
-        <div class="hero-image">
-            <div class="hero-text">
-                <h1>THE BEST MEMORIES ARE MADE AROUND THE TABLE</h1>
-                <p align="center" >Making a reservation at any restaurants in Hoboken city is now easy and takes just a couple of minutes</p>
-                <p align="center" ><a href="/customers/login"  class="fcc-btn">Book a Table</a></p>
-            </div>
-        </div>  
- 
-        <div class="image">
-            <div class="text">
-            <h1 align="center">REACH MILLIONS OF PEOPLE</h1>
-            <p align="center">Add your restaurants infront of millions and earn 3x profits from our listing</p>
-            <p align="center"><a href="/restaurants/register" class="fcc-btn">Add Your Restaurant</a></p>
-        </div>
-        </div>
+<div class="hero-image">
+    <div class="hero-text">
+        <h1>THE BEST MEMORIES ARE MADE AROUND THE TABLE</h1>
+        <p align="center">Making a reservation at any restaurants in Hoboken city is now easy and takes just a couple of
+            minutes</p>
+        <p align="center"><a href="/customers/login" class="fcc-btn">Book a Table</a></p>
+    </div>
+</div>
+
+<div class="image">
+    <div class="text">
+        <h1 align="center">REACH MILLIONS OF PEOPLE</h1>
+        <p align="center">Add your restaurants infront of millions and earn 3x profits from our listing</p>
+        <p align="center"><a href="/restaurants/register" class="fcc-btn">Add Your Restaurant</a></p>
+    </div>
+</div>
