+++ views/layouts/main.handlebars
@@ -6,21 +6,16 @@
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
     <title>{{title}}</title>
     <link rel="stylesheet" href="/public/css/site.css" />
-    <script src="https://ajax.googleapis.com/ajax/libs/jquery/3.5.1/jquery.min.js"></script>
-    <link rel="stylesheet" href="https://maxcdn.bootstrapcdn.com/bootstrap/3.4.1/css/bootstrap.min.css" />
-    <script src="https://maxcdn.bootstrapcdn.com/bootstrap/3.4.1/js/bootstrap.min.js"></script>
 </head>
 
 <body>
 
-        {{{body}}}
-    
+    {{{body}}}
+
     <footer class="footer">
         <p>@ Restaurant Table Reservation Page</p>
     </footer>
-    
 
-    {{!-- <script src="/public/js/form.js"></script> --}}
 </body>
 
 </html>
