+++ views/restaurants/private.handlebars
@@ -1,1 +1,9 @@
-<p>Authenticated</p>
+<section>
+    <nav class="topnav">
+        <a class="active" href="/">Home</a>
+        <a href="/restaurants/profile">Profile</a>
+        <a href="/restaurants/reservation">Reservation</a>
+        <a href="/reviews">Reviews</a>
+    </nav>
+</section>
+
