+++ views/restaurants/register.handlebars
@@ -1,73 +1,91 @@
+<!DOCTYPE html>
+<html>
 
-<section>
-<nav class="topnav">
-<a  href="/">Home</a>
-<a href="/customer/login">Customers</a>
-<a class="active" href="/restaurants/login">Restaurant</a>
-<a href="/reviews">Reviews</a>
-</nav>
-</section>
-<main>
-<div class="jumbotron">
-    <div class="container">
-        <h1>Restaurant Sign Up Page</h1>
-    </div>
-</div>
+<head>
+    <meta charset="UTF-8">
+    <meta name="viewport" content="width=device-width, initial-scale=1.0">
+    <title>{{title}}</title>
+    <link rel="stylesheet" href="/public/css/site.css" />
+    <script src="https://ajax.googleapis.com/ajax/libs/jquery/3.5.1/jquery.min.js"></script>
+    <link rel="stylesheet" href="https://maxcdn.bootstrapcdn.com/bootstrap/3.4.1/css/bootstrap.min.css" />
+    <script src="https://maxcdn.bootstrapcdn.com/bootstrap/3.4.1/js/bootstrap.min.js"></script>
+</head>
 
-<div class="container">
-    <fieldset>
-    <legend>Enter Restaurant Registration Information Here</legend>
-
-    <form class="register" action="/restaurants/register" method="POST">
-        <div class="form-group">
-            <label for="name">Restaurant Name</label>
-            <input type="text" name="name" class="form-control"  placeholder="Enter the Restaurant Name here" required>
+<body>
+    <section>
+        <nav class="topnav">
+            <a href="/">Home</a>
+            <a href="/customer/login">Customers</a>
+            <a class="active" href="/restaurants/login">Restaurant</a>
+            <a href="/reviews">Reviews</a>
+        </nav>
+    </section>
+    <main>
+        <div class="jumbotron">
+            <div class="container">
+                <h1>Restaurant Sign Up Page</h1>
+            </div>
         </div>
 
-        <div class="form-group">
-            <label for="address">Restaurant Address</label>
-            <input type="text" name="address" class="form-control"  placeholder="Enter the Restaurant Address here" required>
-        </div>
+        <div class="container">
+            <fieldset>
+                <legend>Enter Restaurant Registration Information Here</legend>
 
-        <div class="form-group">
-            <label for="email">Email</label>
-            <input type="email" class="form-control"  name="email" placeholder="Email ID:" required>
-        </div>
+                <form class="register" action="/restaurants/register" method="POST">
+                    <div class="form-group">
+                        <label for="name">Restaurant Name</label>
+                        <input type="text" name="name" class="form-control" placeholder="Enter the Restaurant Name here"
+                            required>
+                    </div>
 
-        <div class="form-group">
-            <label for="phone">Contact Number</label>
-            <input type="number" class="form-control" name="phone" placeholder="Enter the Conatact Number here" required>
-        </div>
+                    <div class="form-group">
+                        <label for="address">Restaurant Address</label>
+                        <input type="text" name="address" class="form-control"
+                            placeholder="Enter the Restaurant Address here" required>
+                    </div>
 
-        <div class="form-group">
-            <label for="description">Description</label>
-            <input type="text" class="form-control" name="description" placeholder="Write a short description about the restaurant here" required>
-        </div>
+                    <div class="form-group">
+                        <label for="email">Email</label>
+                        <input type="email" class="form-control" name="email" placeholder="Email ID:" required>
+                    </div>
 
-        <div class="form-group">
-            <label for="passwordHash">Password</label>
-            <input type="password" class="form-control" name="passwordHash" placeholder="Password" required>
-        </div>
+                    <div class="form-group">
+                        <label for="phone">Contact Number</label>
+                        <input type="number" class="form-control" name="phone"
+                            placeholder="Enter the Conatact Number here" required>
+                    </div>
 
-        <div class="form-group">
-            <button type="submit" class="btn btn-primary">Register</button>
-        </div>
-    </form>
-    </fieldset>
+                    <div class="form-group">
+                        <label for="description">Description</label>
+                        <input type="text" class="form-control" name="description"
+                            placeholder="Write a short description about the restaurant here" required>
+                    </div>
 
-    <div>
-        <h3><a href="/restaurants/login">Already a member? Login then!</a></h3>
-    </div>
-</div>
-</main>
-{{#if hasError}}
+                    <div class="form-group">
+                        <label for="passwordHash">Password</label>
+                        <input type="password" class="form-control" name="passwordHash" placeholder="Password" required>
+                    </div>
+
+                    <div class="form-group">
+                        <button type="submit" class="btn btn-primary">Register</button>
+                    </div>
+                </form>
+            </fieldset>
+
+            <div>
+                <h3><a href="/restaurants/login">Already a member? Login then!</a></h3>
+            </div>
+        </div>
+    </main>
+    {{#if hasError}}
     {{#each errors}}
-        <ul>
+    <ul>
         <li class="error">
             {{this}}
         </li>
-        </ul>
-        {{/each}}
-{{/if}}
-        
+    </ul>
+    {{/each}}
+    {{/if}}
+</body>
 
+</html>
