+++ views/users/login.handlebars
@@ -1,23 +1,40 @@
-<section>
+<!DOCTYPE html>
+<html>
+
+<head>
+    <meta charset="UTF-8">
+    <meta name="viewport" content="width=device-width, initial-scale=1.0">
+    <title>{{title}}</title>
+    <link rel="stylesheet" href="/public/css/site.css" />
+    <script src="https://ajax.googleapis.com/ajax/libs/jquery/3.5.1/jquery.min.js"></script>
+    <link rel="stylesheet" href="https://maxcdn.bootstrapcdn.com/bootstrap/3.4.1/css/bootstrap.min.css" />
+    <script src="https://maxcdn.bootstrapcdn.com/bootstrap/3.4.1/js/bootstrap.min.js"></script>
+</head>
+
+<body>
+    <section>
         <nav class="topnav">
             <a href="/">Home</a>
             <a class="active" href="/customers/login">Customers</a>
             <a href="/restaurants/login">Restaurant</a>
             <a href="/reviews">Reviews</a>
         </nav>
-</section>
+    </section>
 
     <div class="container">
         <h1 class="Login">Login</h1>
         <form method="POST" action="/customers/login" id="login-form">
-            <label  for="Email">Email</label>
-            <input class="form-control" type="text" id="Email" name="Email" placeholder="Email" required/> 
-            <label  for="password">Password</label>
-            <input class="form-control" type="password" id="password" name="password" placeholder="Password" required/>
+            <label for="Email">Email</label>
+            <input class="form-control" type="text" id="Email" name="Email" placeholder="Email" required />
+            <label for="password">Password</label>
+            <input class="form-control" type="password" id="password" name="password" placeholder="Password" required />
             <button type="submit" name="submitButton" id="submitButton" class="btn">Login</button>
             {{#each errors}}
-                <div class = "error" id="errors">{{this}}</div>
+            <div class="error" id="errors">{{this}}</div>
             {{/each}}
         </form>
         <a href="/customers/register">Don't have an account? Click Here!</a>
     </div>
+</body>
+
+</html>
