+++ views/users/profile.handlebars
@@ -1,16 +1,16 @@
 <section>
-        <nav class="topnav">
-            <a href="/">Home</a>
-            <a class="active" href="/customers/login">Customers</a>
-            <a href="/restaurants/login">Restaurant</a>
-            <a href="/reviews">Reviews</a>
-        </nav>
+    <nav class="topnav">
+        <a href="/">Home</a>
+        <a class="active" href="/customers/login">Customers</a>
+        <a href="/restaurants/login">Restaurant</a>
+        <a href="/reviews">Reviews</a>
+    </nav>
 </section>
 <div>
     <h1>Hello, {{customer.firstName}}!</p>
-    <br>
-    <p>Email: {{customer.email}}</p>
-    <br>
-    <p>Reviews: {{customer.reviews}}</p>
-    <a href="/customers/logout">Logout</a>
+        <br>
+        <p>Email: {{customer.email}}</p>
+        <br>
+        <p>Reviews: {{customer.reviews}}</p>
+        <a href="/customers/logout">Logout</a>
 </div>
