+++ views/users/signup.handlebars
@@ -1,46 +1,66 @@
-<section>
+<!DOCTYPE html>
+<html>
+
+<head>
+    <meta charset="UTF-8">
+    <meta name="viewport" content="width=device-width, initial-scale=1.0">
+    <title>{{title}}</title>
+    <link rel="stylesheet" href="/public/css/site.css" />
+    <script src="https://ajax.googleapis.com/ajax/libs/jquery/3.5.1/jquery.min.js"></script>
+    <link rel="stylesheet" href="https://maxcdn.bootstrapcdn.com/bootstrap/3.4.1/css/bootstrap.min.css" />
+    <script src="https://maxcdn.bootstrapcdn.com/bootstrap/3.4.1/js/bootstrap.min.js"></script>
+</head>
+
+<body>
+    <section>
         <nav class="topnav">
             <a href="/">Home</a>
             <a class="active" href="/customers/login">Customers</a>
             <a href="/restaurants/login">Restaurant</a>
             <a href="/reviews">Reviews</a>
         </nav>
-</section>
+    </section>
 
-<div class="container ">
-    <h1 class="Signup">Sign Up</h1>
-    <form method="POST" action="/customers/register" id="signup-form">
+    <div class="container ">
+        <h1 class="Signup">Sign Up</h1>
+        <form method="POST" action="/customers/register" id="signup-form">
             <div class="form-group">
                 <label for="firstName">Enter the FirstName</label>
-                <input type="text" class="form-control" id="firstName" name="firstName" placeholder="First Name"required/>
+                <input type="text" class="form-control" id="firstName" name="firstName" placeholder="First Name"
+                    required />
             </div>
             <div class="form-group">
                 <label for="lastName">Enter the LastName</label>
-                <input type="text" class="form-control" id="lastName" name="lastName" placeholder="Last Name" required/>
+                <input type="text" class="form-control" id="lastName" name="lastName" placeholder="Last Name"
+                    required />
             </div>
             <div class="form-group">
                 <label for="email">Enter the Email</label>
-                <input type="email" class="form-control" id="email" name="email" placeholder="Email Address" required/>
+                <input type="email" class="form-control" id="email" name="email" placeholder="Email Address" required />
             </div>
             <div class="form-group">
                 <label for=" phone">Enter the PhoneNumber</label>
-                <input type="number" class="form-control" id="phone" name="phone" placeholder="PhoneNumber" required/>
+                <input type="number" class="form-control" id="phone" name="phone" placeholder="PhoneNumber" required />
             </div>
             <div class="form-group">
                 <label for="city">Enter the City</label>
-                <input class="form-control" id="city" name="city" type="text" placeholder="City" required/>
+                <input class="form-control" id="city" name="city" type="text" placeholder="City" required />
             </div>
             <div class="form-group">
                 <label for="state">Enter the State</label>
-                <input type="text" class="form-control" id="state" name="state" placeholder="State" required/>
+                <input type="text" class="form-control" id="state" name="state" placeholder="State" required />
             </div>
-             <div class="form-group">
-                <label for="password" >Enter the Password</label>
-                <input class="form-control" id="password" name="password" type="password" placeholder="Password" required/>
+            <div class="form-group">
+                <label for="password">Enter the Password</label>
+                <input class="form-control" id="password" name="password" type="password" placeholder="Password"
+                    required />
             </div>
-        <button class="btn" type="submit" name="submitButton" id="submitButton">Signup</button>
-        {{#each errors}}
-            <div class = "error">{{this}}</div>
-        {{/each}}
-    </form>
-</div>
+            <button class="btn" type="submit" name="submitButton" id="submitButton">Signup</button>
+            {{#each errors}}
+            <div class="error">{{this}}</div>
+            {{/each}}
+        </form>
+    </div>
+</body>
+
+</html>
